requirements.txt
flask==1.1.2
flask-restx==0.2.0
pyyaml==5.3.1

numpy==1.16.3
opencv-contrib-python==3.4.2.17
scipy==1.3.0
